app/requirements.txt
passlib[bcrypt]
email-validator
fastapi
python-jose[cryptography]
python-multipart
